requirements.txt
asgiref==3.9.1
autoflake==2.3.1
black==25.1.0
cfgv==3.4.0
click==8.3.0
colorama==0.4.6
distlib==0.4.0
Django==5.2.6
django-bootstrap5==25.2
django-filter==25.1
django-htmx==1.25.0
filelock==3.19.1
flake8==7.3.0
identify==2.6.14
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.1.0
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
pyzbar==0.1.9 # на мак https://github.com/npinchot/zbar/issues/3#issuecomment-1038005495
